# cmd/stationd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "drop api document.",
        "title": "drop API",
        "contact": {
            "name": "tk",
            "email": "tpkeeper.me@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/feeStation/api",
    "paths": {
        "/v1/station/poolInfo": {
            "get": {
                "description": "get pool info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "get pool info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Rsp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/station_handlers.RspPoolInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/station/swapInfo": {
            "get": {
                "description": "get swap info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "get swap info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "block hash hex string",
                        "name": "blockHash",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tx hash hex string",
                        "name": "txHash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Rsp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/station_handlers.RspSwapInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "post swap info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "post swap info",
                "parameters": [
                    {
                        "description": "user swap info",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/station_handlers.ReqSwapInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Rsp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "station_handlers.PoolInfo": {
            "type": "object",
            "properties": {
                "poolAddress": {
                    "description": "base58 or hex",
                    "type": "string"
                },
                "swapRate": {
                    "description": "decimals 6",
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "station_handlers.ReqSwapInfo": {
            "type": "object",
            "properties": {
                "blockHash": {
                    "description": "hex",
                    "type": "string"
                },
                "inAmount": {
                    "description": "decimal",
                    "type": "string"
                },
                "minOutAmount": {
                    "description": "decimal",
                    "type": "string"
                },
                "poolAddress": {
                    "type": "string"
                },
                "pubkey": {
                    "description": "hex",
                    "type": "string"
                },
                "signature": {
                    "description": "hex",
                    "type": "string"
                },
                "stafiAddress": {
                    "description": "hex",
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "txHash": {
                    "description": "hex",
                    "type": "string"
                }
            }
        },
        "station_handlers.RspPoolInfo": {
            "type": "object",
            "properties": {
                "poolInfoList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/station_handlers.PoolInfo"
                    }
                },
                "swapMaxLimit": {
                    "description": "decimals 12",
                    "type": "string"
                },
                "swapMinLimit": {
                    "description": "decimals 12",
                    "type": "string"
                }
            }
        },
        "station_handlers.RspSwapInfo": {
            "type": "object",
            "properties": {
                "swapStatus": {
                    "type": "integer"
                }
            }
        },
        "utils.Rsp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}